pyproject.toml
[tool.poetry]
name = "reservationbackend"
version = "0.1.0"
description = "Backend for a reservation app for demostrate the knowledge acquired in the ioet university 2"
authors = ["Gerardo Rios"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pytest = "^8.2.2"
neo4j = "^5.22.0"
mypy = "^1.10.1"
flake8 = "^7.1.0"
black = "^24.4.2"
uvicorn = "^0.30.1"
pre-commit = "^3.7.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
pymysql = "^1.1.1"
cerberus = "^1.3.5"
pytest-mock = "^3.14.0"
bcrypt = "^4.2.0"
alembic = "^1.13.2"
isort = "^5.13.2"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
